Batch operations complete when the result receivers complete (not only finished)
We encountered some issues where when executing batch updates using the
postgres protocol, the returned row count for the operation would be 0.
This happened because when the operation is complete we signal that the
backend is read for query (using the readyForQuery message).
We used to consider the operation complete after we sent the "commandComplete"
messages for every batch in the operation.
Due to the commandComplete channel transimission being async, sometimes
we sent readyForQuery before commandComplete would actually be sent.

This commit changes the definition of "operation complete" from "all result
receivers finished" to "all result receivers complete" (ie. commandComplete
was delivered)